Image improvements
Spent ages fiddling with the images trying out directives, polymer iron
image and then just setting styles myself. Learnt background image css
is the best for fitting to elements, iron-img had issues retaining
ng-style properties along with it's sizing parameter, and angular
directives seemed to lose some of their styles occasionally in the
sidebar for no apparent reason... In the end, 2 div's and a small amount
of css is the best and easiest solution.

This changes prepares images for the image selector feature to be added,
and for mobile features to be continued on.